+++ app/src/main/java/com/example/playlistmaker/SearchActivityAPI.kt
@@ -1,0 +1,12 @@
+package com.example.playlistmaker
+
+import retrofit2.Call
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface SearchActivityAPI {
+    @GET("/search?entity=song")
+    fun getTracks(
+        @Query("term") text: String
+    ): Call<TracksResponse>
+}
